macOS: Fix opening LilyPond files from the Finder
On Linux and Windows, opening a LilyPond file with Frescobaldi from the
file explorer while no Frescobaldi instance is running launches the
`frescobaldi` executable with the file as a command-line argument. On
macOS, the mechanism is different: the system sends an event to the
application, which Qt translates to a QFileOpenEvent.

Frescobaldi has a filter catching this event, but it was being added too
late. The QFileOpenEvent's are sent by Qt when the event loop is first
started by the splash screen, and the filter was only installed
afterwards. As a result, double-clicking on a LilyPond file on the
Finder without an already running Frescobaldi instance was launching but
not actually opening the file, whereas doing it whith an existing
Frescobaldi instance did open it.

To solve this, we need to differentiate between events we get during
startup and other events.